+++ TRRandomizerCore/Randomizers/TR3/Classic/TR3EnvironmentRandomizer.cs
@@ -5,6 +5,7 @@ using TRLevelControl.Model;
 using TRRandomizerCore.Helpers;
 using TRRandomizerCore.Levels;
 using TRRandomizerCore.Textures;
+using TRRandomizerCore.Utilities;
 
 namespace TRRandomizerCore.Randomizers;
 
@@ -192,35 +193,6 @@ public class TR3EnvironmentRandomizer : BaseTR3Randomizer, IMirrorControl
             monitor.UseMirroring = true;
         }
 
-        CheckMonkeyPickups(level);
-    }
-
-    private static void CheckMonkeyPickups(TR3CombinedLevel level)
-    {
-        // Do a global check for monkeys that may be sitting on more than one pickup.
-        // This has to happen after item, enemy and environment rando to account for
-        // any shifted, converted and added items.
-        List<TR3Entity> monkeys = level.Data.Entities.FindAll(e => e.TypeID == TR3Type.Monkey);
-        foreach (TR3Entity monkey in monkeys)
-        {
-            List<TR3Entity> pickups = level.Data.Entities.FindAll(e =>
-                    e.X == monkey.X &&
-                    e.Y == monkey.Y &&
-                    e.Z == monkey.Z &&
-                    TR3TypeUtilities.IsAnyPickupType(e.TypeID)).ToList();
-
-            if (pickups.Count == 1)
-            {
-                continue;
-            }
-
-            // Leave one item to drop, favouring key items. The others will be shifted
-            // slightly so the monkey doesn't pick them up.
-            pickups.Sort((e1, e2) => TR3TypeUtilities.IsKeyItemType(e1.TypeID) ? 1 : -1);
-            for (int i = 0; i < pickups.Count - 1; i++)
-            {
-                ++pickups[i].X;
-            }
-        }
+        TR3EnemyUtilities.CheckMonkeyPickups(level.Data, false);
     }
 }